refactor: split file tree tests by class
